backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pydantic==2.4.2
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.1
